re-added tests, but testing resolve options per discussion in PR review
